+++ Library/Aphid/UI/Controls.js
@@ -1,0 +1,7 @@
+/**
+ * Aphid.UI.Controls
+**/
+
+Aphid.UI.Controls = {};
+
+//= require "Controls/SelectView"